# Speed up `test-decks_test.py`