--- a/test-decks/errors/bad-crop.deck
+++ b/test-decks/errors/bad-crop.deck
@@ -1,4 +1,4 @@
-#ends: Error in ffmpeg. See above.
+# Error running ffmpeg
 title: Test deck::Errors::Bad crop
 crop: 100000:1