--- a/test-decks/good/slow.deck
+++ b/test-decks/good/slow.deck
@@ -1,8 +1,6 @@
 title: Test deck::Partial slow
 note_id: {deck_id} {text}
 
-file://./media/stopwatch.mp4 BASE
-
 file://./media/stopwatch.mp4
   slow: - * 0.1
   without trim: slow: - * 0.1
@@ -20,13 +18,3 @@ file://./media/stopwatch.mp4
 file://./media/stopwatch.mp4
   slow: 0.5- * 0.25
   with 1s-2s trim: slow: 0.5- * 0.25
-
-file://./media/stopwatch.mp4
-  video: strip
-  slow: 0.5- * 0.25
-  no video / with 1s-2s trim: slow: 0.5- * 0.25
-
-file://./media/stopwatch.mp4
-  audio: strip
-  slow: 0.5- * 0.25
-  no audio / with 1s-2s trim: slow: 0.5- * 0.25