--- a/test-decks/good/tags.deck
+++ b/test-decks/good/tags.deck
@@ -2,11 +2,11 @@ title: Test deck::Tags
 tags: test::a test::b
 note_id: {deck_id} {text}
 
-file://./media/stopwatch.mp4 Tags [test::a, test::c]
+file://./media/first.png Tags [test::a, test::c]
   tags: +test::c -test::b
 
-file://./media/stopwatch.mp4 Tags [test::a, test::b]
+file://./media/first.png Tags [test::a, test::b]
 
 # Remove tag that doesn’t exist
-file://./media/stopwatch.mp4 Tags []
+file://./media/first.png Tags []
   tags: -test::a -test::b -test::c