- **Faster parsing in `test-decks_test.py`**
  This switches back to parsing the decks in the same process rather than
  calling out to a subprocess. On my laptop, the test goes from a mean
  time of 11.0 seconds to 6.4 seconds.
  

- **Speed up decks in `test-decks`**
  Trim some excess notes in the decks, and switch media to still images
  where possible. This improves `test-decks_test.py` time from 6.4 seconds
  to 4.1 seconds.