@danielparks 2025-01-18
Most of the time in CI checks is spent installing `ffmpeg`.